OSS-fuzz #43264 - do not confuse images and masks
The code to fill in the image information for rendering could end up
not setting the colour space, if 'ImageMask' is true (these are
specified as incompatible). This could lead to us processing an image
as an image (not a mask) but not having a colour space to use for it,
leading to a crash.

Check BPC and if it is 1, then treat as a mask, otherwise treat as an
image. Flag an error.