OSS-fuzz #43354 - Validate Default* colour spaces
We were not checking that a DefaultGray, DefaultRGB or DefaultCMYK
space was not one of Lab, Pattern or Indexed nor that it had the correct
number of components.

This commit checks that, moves the loading of Default* colour spaces
into pdf_colour.c instead of doing almost exactly the same complex task
in two different places (neither of them the colour code) and if the
space is illegal, raies a warning.