GhostPDF - Implement -dNOTRANSPARENCY
This is slightly different (I think) to the old scheme. It turns out
that the transparency implementation is contingent on the result of the
checking done on each page. The code check ctx->page.has_transparency
to see if it should implement transparency.

Rather than add an additional test in all the places where this is
checked it is simpler to test the setting of NOTRANSPARENCY in the
per-page spot colour/transparency checking code and not set the page
'has_transparency' flag if -dNOTRANSPARENCY is set.