OSS-fuzz #43268.pdf - beware of repair rebuilding xref
The file is broken (of course) and causes a repair during the course
of reading an object form the file while dereferencing an indirect
reference.

The repair causes the xref table to be discarded and rebuilt, and that
causes the reference to the xref entry to become invalid. Simply update
the reference after reading an object to ensure it is up to date.