oss-fuzz 42751: Failures in pdf14 and in image color setup
If the page has a pdf14 compositor device and we fail in the
processing of the page, then abort the pdf14 device rather than
popping the device.

Also, if gx_set_dev_color fails when using a profile embedded in
an image and we try to use a default colorspace in its place,
try gx_set_dev_color again.  If that fails (likely due to memory
issues), we need to exit.