oss-fuzz 42998: Drawing text: don't rely on the stack reference
for the string operand.

We were relying on the operand stack reference for the string object to remain
valid for the duration of the text drawing operation, but this is not sure to
be the case. If a pdfi error occurs during the text drawing it can cause the
entire stack to be cleared. Thus we need to a "local", counted reference to the
string, so the object remains valid until we're finished with it.