OSS-fuzz #43247 - special treatment for reading streams
When reading objects from the PDF file we don't generally dereference
anything, we just store indirect references as indirect references.
This insulates us from circular references in obejcts at this point
and the loop detection machinery caters for later dereferencing.

However, streams are an exception, we validate the /Length when we read
the stream object, which means that if the /Length is associated with
an indirect reference we will try to dereference it. If the reference
points back to the same stream object, we end up in infinite recursion.

Do special case loop detection just for this.