Memento: Fix operation with valgrind.
Recently, I updated memento to use a splay tree for the
location heap. This made it work much faster. Unfortunately,
this broke operation with valgrind, as the block headers
(in which the pointers for the tree are held) are marked
as 'UNDEFINED' memory most of the time (to allow valgrind
to accurately detect over/underruns).

The fix is to add a load of valgrind macros to memento,
done here.