Fix PCL interpreter leaking a graphics state.
A memento build of pcl, run with no arguments shows significant
leakage. Part of this is due to a graphics state leaking from the
PCL interpreter.

This appears to be because of unexpected (to me at least!) behaviour
with gsave/grestore. If we make a graphics state, and then gsave,
we might reasonably expect grestore to leave us back in the newly
created state.

This is not the case; when grestore restores, if it would leave us
back at the 'topmost' gstate, it does an extra gsave. This is
believed to be to do with coping with Postscript files that have
more grestores than gsaves.

Accordingly, we work around it here by doing an extra gs_gsave at
the start, and an extra gs_restore_only at the end.

This solves at least some of the leaks.