OSS-Fuzz 43253.pdf: clist bug
Thanks to Ken for basically doing all the investigation on this.

In the clist code cmd_write_buffer can return 1 to 'warn' of an
approaching low memory problem. The call to this from cmd_put_range_op
was failing to detect this, due to checking for "!= 0" on return
rather than the more usual "< 0".

Simply fixing this solves the bug.