oss-fuzz 42556 ICC cache clean-up
When a failure occurs during the creation of an ICC
link that is getting added to the link cache, we
want to make sure to release and remove the link
object that was added to the cache.

Also add a max count in the while loop that is
continually releasing the link lock waiting for
another thread to make it valid.  Some other issues
(like the one above) can cause us to get stuck
in this loop. If this does happen we should look
closely at why it is happening, as it should not.
This fix throws a warning and avoids the endless
loop.