OSS-fuzz #43595 - more checks on image/mask
Another broken image which has both a ColorSpace and ImageMask entry
(mutually exclusive). In this case BPC is 1 so we were treating it as
a mask for the purposes of extracting the colour space, but then later
in the image processing treating it as an image (because is has a /Mask
as well).

Extend the test for assigning a colour space to cater for this too.