Don't forcibly replace glyph metrics in Freetype
Previously, Freetype was unclear about the incremental interface's call back
to retrieve glyph metrics. It implied it was a way to supply "custom" metrics.

That turns out not to be the case: it is intended for the (extremely rare!)
cases when the "real" glyph metrics are not available directly from the the
glyph description, font tables etc. Thus using it for custom metrics caused
problems for Truetype glyph hinting programs, because those metrics are
used to populate the zone points which the hinting programs use.

Freetype resolved things at their end, and I put a solution in place for the
interface code, but it turned out to be insufficient. It was still allowing the
custom advance width to be set.

This caused glyph hinting issues, for example:
tests_private/comparefiles/Bug691824.ps

at 600dpi.

And spacing issues, with many PCL tests, for example:
tests_private/xl/xlats/WWEPGUJ3.BIN

at 600 dpi.

This prevents *any* custom metrics being supplied to Freetype in cases where it
causes problems (mainly incremementally defined TTFs), and handles custom
advance widths outside of Freetype.