oss-fuzz 42920: don't set object type to NULL
Allegedly, to avoid running the device "finalize" method when cleaning up after
an out-of-memory error, the code was setting the object type in the memory
header to NULL (instead of its original value of st_device_memory. But none of
the memory managers cope with the type being NULL.

Set the object type to st_bytes (instead of NULL) that way the memory is
treated as a simple byte array (i.e. no finalize method) in a way the memory
managers understand.