OSS-fuzz #43621 - prevent overflow in pattern step scaling
Robin Watts also has a fix for this, in the graphics library, I think
it is worth having both.

The problem was that YStep in the pattern was so large that when scaled
by the CTM it was overflowing a float and becoming #INF, which was
making all the arithmetic in the graphics library invalid.

This code checks the Step values to ensure they all fit into
a float, and then in the graphics library checks that after
multiplication they still fit before attempting to use them.