GhostPDF - Fix PostScript cleanup on error in showpage
Michael ran across this under a different set of conditions. For me

-K1048576 -r200x200 -sBandListStorage=memory -dMaxBitmap=0 -dBufferSpace=450k -dMediaPosition=1 -dcupsColorSpace=1 -dSAFER -dNOPAUSE -dBATCH -dNOINTERPOLATE -dNOMEDIAATTRS -sOutputFile=/temp/out.tif -sDEVICE=tiff24nc /bugs/oss-fuzz/42920.pdf

Using the file from OSS-fuzz bug 42920.pdf, with a 64-bit debug build
on Windows led to an error on closing the interpreter because we were
finalizing the input file twice. I'm not entirely certain why, but
putting the showpage in a stopped context so that if it fails we carry
on and clean up properly resolves it for me.