OSS-fuzz #42995
We were keeping a pointer to the Trailer dictionary while reading an
XRefStm object. But it is possible for that to repair the PDF file and
the action of repairing the PDF file can cause ctx->Trailer to be
discarded and rebuilt, leaving us with a dangling pointer.

Make sure that we refresh the pointer if that happens.