Remove no longer required font object reference counting
When switch between Postscript and pdfi, we previously had to handle font
object reference counting explicitly, but now the current font is managed with
the pdfi graphics state, it's no longer required.

It didn't cause problems because pdfi starts with no current font, and works
hard to ensure it finishes with the gstate in the same state it started.