Fix SEGV seen in alldevs test.
  gpcl6 --debug=epo-disable -sOutputFile=out%04d.tif -dTIFFDateTime=false
  -dMaxBitmap=400000000 -sDEVICE=tiffsep1 -r72 -Z: -dNOPAUSE -dBATCH
  -K2000000 -dClusterJob ../perf-testing-gpdl/pcl/j9_acrobat.pcl

was crashing. This is because pxBeginPage was putting a parameter into
the device which was causing it to close. The device was then not being
opened again before it was being used.

The fix is simply to make the code reopen the device after all the
put_paramming has completed.