Check the return values from setting up the Default colour and QState
When running a stream (XObject, type 3 character etc) we may need to
set up the DefaultGray etc colour spaces and we may need to set up
a graphics state which differs from the current one.

We were not checking the return code from the functions performing
these tasks.

In this case, if we cannot set up the required state, we simply abandon
the stream (it probably wouldn't be correct anyway)