Fix gslp.ps with -p option
Commit 8d8e90560bf6bbf0415b28445e7d49c005454f77 added a new option
'encoding' which is intended to specify the font encoding to be used
instead of always using ISOLatin1.

However, the commit broke the output PostScript produced by use of the
-p option, and didn't re-encode the default fonts when -p was used,
which meant the re-encoding didn't work with the default fonts, only if
a specific font was used (-f command line option).

This fixes both problems, it's not as neat as the original code, but
the previous commit means that ship has already sailed.