oss-fuzz 42612: Only draw text with a valid font type
In this case, the file sets a Type 11 (CIDFontType 2) font, without the required
Type 0 "parent". That causes the pdfi text enumerator setup code to skip the
setup of the byte buffer representing the text to be drawn, but it was still
attempting to draw the text.

Change it so attempting to draw text with an invalid font type throws an error.