OSS-fuzz #43914 - Clear Mask entry on broken ImageMask
If we have a confused Image/Mask dictionary (elements of each present)
and we choose to treat it as an image, make certain that any existing
Mask entry is counted down and released.

Otherwise the image code will attempt to treat the mask as an image,
and tries to use a NULL colour space.