Don't countdown a non-existent pattern
One error case where we're handling a pattern, but the error occurs before the
pattern object is created. In that case, we shouldn't attempt to decrement the
reference count of the local pattern reference.

Noticed investigating oss-fuzz issue: 43991 but the error location doesn't
match what oss-fuzz reports, so I'm not sure it's the same probem.