OSS Fuzz 42916: Fix memory corruption due to dynamic_grow not growing.
The current version of dynamic grow can end up not growing a buffer.

Consider an old_size of 20, a new_size of 40. If this fails, we retry
with 30, then 25, then 22, then 21, then 20, which is no change, so
succeeds.

Move the test for new_size <= old_size loop termination into the loop
and everything works as it should.