Fix memory squeezing leak in gdev_prn_allocate.
Found when squeezing the file from OSS Fuzz 42916, but this is
probably independent of that file. Ensure we free the bg_print
allocation if the prn_allocation fails.

I do wonder about maybe cleaning other structures up if the
open_device fails, but we'll revisit that after some more
squeezing.