OSS-fuzz #43917 - don't try to countdown a non-object
We get the offset of the start of the xref as an integer, not a PDF
object. So if we fail and exit don't try to countdown the start.

This only happened in one place, I suspect we changed the code from
using a PDF object to an integer and missed this cleanup case.