base/memento.c: limit output of detailed blocks, and optionally abort after corruption.
If memento.showDetailedBlocks is positive, showInfo() now decrements it
and returns 1 immediately if it reaches zero. So for example setting
MEMENTO_SHOW_DETAILED_BLOCKS=10 limits the max number of detailed block
diagostics to 10. showDetailedBlocks now defaults to -1 which shows all blocks
as before.

In Memento_Internal_checkAllAlloced(), call abort() if we find corruption and
memento.abortOnCorruption is true.

Also fixed warning from clang and gcc for initialisation of Memento_hashedST's
trace[1] array.