Fix memory leak.
When running a windows memento, 32bit build of:

 gpcl6 -dNOPAUSE -dBATCH -sDEVICE=ppmraw -dMaxBitmap=1000 -r300
    -o gpcl6.out ../tests_private/pcl/pcl5cats/*.BIN

we'd see leaks from pl_bitmap_build_char. I could not reproduce
these when running individual files.

A desk-check of the code spotted that penum would not be freed if
a particular error path was taken. Fixing that (as here), removes
the leaks.