Revise FAPI integration in PCL code.
Originally, the FAPI code held a buffer of the full font data, for the duration
of the life of the font object.

Given that the entire font set is memory resident, three times over, for the
life of the interpreter(s), that ends up holding quite a lot of memory.

This changes it so it uses the font data read directly from the romfs (rather
than read into a persistent memory buffer). To maintain performance, we now
store the PCL fonts uncompressed in the romfs. This grows the binary from
25Mb to 29Mb.