OSS-fuzz #44232 - sort loop detection in pdfmark generation
The pdfmark generation code ignores circular references in dereferenced
objects, which I'm not certain is going to be viable in the long term,
but I've chosen to leave alone just now.

However, it was not checking for loops when writing dictionary contents
which could, with a sufficiently broken annotation, cause us to enter
an endless loop trying to write a dictionary with looping references.

This addresses the problem, but I suspect more work needs to be done
here.