OSS-fuzz #44258 - treat /Prev XRefStm with increased /Size as an error
The file has an XrefStm which has a /Prev in the stream dictionary. The
/Prev key has a value pointing to a previous XRefStm (so this is an
incrementally updated file).

The /Size in the initial XRefStm should be at least as large as the
/Size in the previous XRefStm (it is the largest object in the PDF file
and incrementally updated files do not remove objects). In this case
the previous XRefStm has a larger /Size than the initial one, which is
illegal.

We were then attempting to read more entries into the xref table than
we had allocated, leading to writing off the end of the buffer.

Fix by treating an increased /Size in a /Prev xref as an error.