Rework MSVC makefiles slightly.
Currently, we are building the .exe files with one crt library
and the dll with another. The only downside to this appears to
be that if you set a file descriptor to be 'binary' (for instance
stdin) in the app, the dll will still see it as text. This
causes corruption of files when they are catted in via stdin.

The fix here is to build all files in the same way. If MAKEDLL is
used, everything uses the DLL based crt, if not, nothing does.

This causes changes in which libraries are used. Listing such
libraries in the makefile becomes problematic due to VS2015
and onwards having changed the libraries in use.

Fortunately, because we are now using the correct (and consistent)
flags to the compiler, the default libraries are now correct,
which means we can now avoid having to mention the libraries by
name at all.

This has been tested with both VS2005 and VS2019.