Coverity ID 375068 - fix uninitialised variable usage
The code had two initialisations which need to be performed before any
error handling in the 'exit' goto, which meant if the first failed then
we could try to use the second uninitialised.

There's no way to initialise both, but we can avoid the problem by
simply returning immediately if the first initialisation fails, instead
of jumping to the exit.