oss-fuzz 44524: Leave path in a valid state, despite error
In the event of an error allocating new path segments, we could end up leaving
the "segments" element set to NULL, a condition the rest of the path code
does not expect to happen.

Rather than slow things down by checking segments != NULL on every path
operation, put the old segments back in place before returning the error,
thus meaning although the contents are "wrong" the path is always in valid
state.