Coverity fix: Increase declared size for pdf_name etc.
Previously, we'd defined pdf_name/pdf_string/pdf_keyword as
having sizeof(int) sized data arrays, and actually alloced
them to be larger. This upsets coverity as it can infer that
we're accessing beyond the end of the structure.

So, we up those declared sizes to be 4096 bytes, and actually
allocate them to be smaller as required. This should hopefully
be enough to satisfy coverity's checking, without actually
changing the amount of memory we use.