GhostPDF - Use the ICC cache for ICCBased spaces
This is a performance enhancement, when we are asked to create a new
ICCBased colour space, we now probe the ICC cache, using the object
number of the ICC colour space as a key, to see if it has previously
been created and cached.

If we find the colour has already been cached, then use the cached copy
instead of reading the ICC profile. This massively improves performance
for file which do the equivalent of:

q
/CS1 cs
...
Q
q
/CS1 cs
...
Q

We already have a check for named colour spaces which are the same as
the current space, but the use of gsave/grestore was defeating that.

The QL ATS 1.7 test file CDX3C1DC_Dist.pdf runs about 30 times faster
with this change. Other files are also affected.

This also fixes (I have no idea why)
eci_altona-test-suite-v2_technical2_x4.pdf with the psdcmyk16 device.

Thanks to Michael for the vital clue about the ICC override when we are
processing a soft mask.