Cleanup memory on exit from shading with error
File \tests_private\pdf\customer394\problems\normal_990.pdf, with
the display or pkmraw devices, hits an error in this shading. The
shading function then exits, but fails to clean up the patch_fill_state
leading to it leaking colours and vertex edges.

This only occurs with pdfi since the garbage collector catches
the leaked memory and frees it with Ghostscript.