GhostPDF - fix a memory leak with named colour spaces
Since we store the name of a colour space in the interpreter_data when
we have a named space, and count up the name, we must not arbitrarily
replace it when we call pdfi_set_colour_callback().

This could happen if we reused a space, eg a cached ICCBased space,
because when we set the colour space in the graphics state we call
pdfi_set_colour_callback().