Softmask object memory leak. Related to Bug 693312
When the source file has already created a softmask
and is using another we currently destroy the mask
stack and create a new one, which was related to
Bug 693312.  Unfortunately, we were not ensuring
that the masks in the stack were getting
released.  See

\tests_private\comparefiles\Bug691803.pdf
-o output_cmyk.tif -sDEVICE=tiff32nc Bug691803.pdf

for an example memory leak in gpdf with memento build.