GhostPDF - memory leak caused by using wrong dictionary
The code here used the original page dictionary when it should have
been using the Action dictionary. The end result of this is that we
ended up with page dictionaries pointing to annotations which contained
references back to page dictionaries.

Although all the reference counts were correct, the circular referencing
meant we never freed the mutually referencing objects, and so leaked
memory. Since one of the objects was a page dictionary, this meant we
leaked a lot of memory.

Fixed by using the correct dictionary. Noted also that we don't
need to retrieve the Action dictionary when we dereference it, we aren't
going to use it again so it's safe to leave it as NULL.