Coverity 375583: Check the size of the file mode string
Slightly speculative fix for this CID.

The string for file access modes should never be more than 3 bytes: r, r+, w,
w+, a, a+, and a NULL terminator. Internally, on platforms that require it,
we automatically append a further mode for "binary", so the total size,
including terminator, should never be more than 4 bytes.

Ensure the initial string we're handed is 2 bytes or less.