Bug 704945: Add init_device_procs entry for mem_x_device.
When allocating a buffer device, we rely on an init_device_procs
being defined for the device we are using as a prototype. Which
device we use as a prototype depends upon the number of bits per
pixel we are using. For bpp > 64, we use mem_x_device, which does
not currently have an init_device_procs defined.

This is a fairly hard case to tickle, as very few devices use
more than 64 bits per pixel. The DeviceN device is one of the
few that does, and then the problem only kicks in if the
MaxBitmap figure is high enough (or conversely the resolution is
low enough).