pdfi: add option to not replace existing entries in dictionaries
In certain cases, we want to ignore duplicate keys being written to a
dictionary, so add a parameter to pdfi_dict_put_obj() telling it to behave
thus.

It saves the expense of doing a pdfi_dict_known_by_key() first.

And take advantage of it in the Type 1 font code.