GhostPDL - performance improvement with large dictionaries
The file \tests_private\pdf\PDF_1.7_ATS\CDX3C1DC_Dist.pdf has a page
with a /Resources dictionary which has an XObject dictionary with over
10,000 key/value pairs.

We need to retrieve each of those at least once in order to draw every
image, and the time taken to find the correct key was comparatively
large because we were doing a simple linear search.

This commit adds a new 'is_sorted' property to the dictionary, if the
dictionary is sorted then we use a binary search to find the required
key, otherwise we use a linear search.

We sort the dictionary when a key is retrieved or it's existence is
checked and we only do this for larger (> 32) dictionaries. We defer the
sorting till key retrieval as this avoids us continually resorting
the dictionary while it is being built. In general we don't modify
dictionaries often after creation with PDF (unlike PostScript).

We had hoped this wouldn't be required, but the test file shows a 25%
performance improvement with this fix, putting it back more or less the
same as GS (which has a complex hashing mechanism for dictionaries).