GhostPDF - handle encrypted file with both StmF and StrF /Identity
The file /tests_private/pdf/PDF_1.7_ATS/AC8ZD1CC_Save.pdf is an
encrypted file, using security handler version 4, PDF 1.5, where the
Encrypt dictionary has both the StmF and StrF filters set to /Identity.

This means the file is, in effect, not encrypted. Though technically it
is possible for a compressed object stream to specify a different
filter.

We were unable to create an object key because both filters were
Identity, leading to us failing to open the file. This commit will
continue in that case, and also will short-circuit a little of the
setup which is not required if the filter being used is Identity for
either streams or strings.

Previously the code worked correctly, it just went through some hoops
that we don't actually need to do, so this is (very) slightly more
efficient.

Ghostscript gives a warning about this case, but I've chosen not to here
as technically the file is entirely valid from that point of view (it
has many other problems). If anyone ever comlains we can add the
warning easily enough.