OSS_fuzz #44852 - copy/paste error in pdfi_o_BDC
The object type check for the properties was checking the wrong object
(almost certainly a copy/paste from the top of function). Not a problem
for well-formed files, but the test file had an integer instead of a
name, trying to access the 'data' property of the name led to a read
from random memory.