GhostPDF - Performance enhancement don't cache type 3 sub-functions
The test file /tests_private/pdf/pdf_1.7_ATS/WWTW61EC_file.pdf exhibits
a marked performance problem on page 2.

This is because it has many shadings, each of which use one of a small
number of functions. We cache the functions, as with all objects, so
would expect these to be processed efficiently. However, the functions
are type 3 stitching functions and each one has 255 sub-functions.

This means that as we process the function array we eject every cached
item (and then some of the sub-functions), which means that we lose the
benefit of the cache.

This commit adds 'nocache' options to the derefrencing code, and new
options in array and dictionary handling to retrieve objects without
adding them to the cache. We only actually use the nocaching array
routine, but it seems worth having a matching dictionary version
available.

This improves the performance of the file in question by a factor > 10.