Bug 703308: Revise X11 devices to honour HWResolution
It was observed some time ago that some X11 implementations would crash or
(effectively) freeze if a sufficiently large area window was requested.

The original solution was to refuse requests to change the resolution, and clamp
the media size to the available screen area. The resulting behaviour means
some Postscript programs don't behave the manner one might expect.

This changes the code to allow changes in HWResolution then, again, clamps the
media size to the available screen area.

It still leaves (what I consider) slightly unexpected behaviour when we are
forced to clamp the values.