GhostPDF - Fix Outlines processing
In a fir of mis-naming by Adobe, the /Count parameter of an Outline
does not tell you how many descendants a given outline node has, it
tells you how many open items there are at all levels for the initial
node (or missing if there are no open outline items). For descendant
nodes this is either the number of open outlines at all levels below
this one (if the item is open) or the number of descendants which would
appear if the item were opened if it is closed, the latter expressed as
a negative value.

The code was counting the total number of descendant items and using
that as the /Count value, which is simply wrong. We cannot realistically
determine what the correct value should be, we just have to accept
whatever was sent to us.

So this is just a case of deleting lots of code.