Fix segfault issues with simulate overprint
When simulate overprint is set for a sep device, there is no
need to do a simulation as the device already supports spot
color overprinting.  This will change in the future for cases
dealing with rendering to the output intent but the target
device uses a different ICC profile.  This issue occurred only
with pdfi.

There was also a problem with the use of patterns in the
case where we are doing overprint simulation. If the pattern
accumulator target is the pdf14 device and we are doing
overprint simulation, then it is necessary
that the pattern accumulator uses transparency.  This problem
occurs for the old PDF interpreter and pdfi.  This commit
fixes the issue for pdfi.  The old PDF interpreter needs an
additional fix.

Example command line with issues
-sDEVICE=bitrgbtags  -dMaxBitmap=10000000000  -dNEWPDF=true/false -dOverprint=/simulate  -r72 -o testout.ppm -f Bug693541.pdf